Create Table of Risk Estimates

Creates a table showing the estimated risk that one or more people will be infectious for the given input locations, event sizes and ascertainment biases.

create_c19r_data(
  df_in,
  risk_output = sprintf("world_risk_regions_%s.csv",
    stringr::str_replace_all(lubridate::today(), "-", "")),
  output_prefix = ".",
  event_size = c(10, 15, 20, 25, 50, 100, 500, 1000, 5000),
  asc_bias_list = cbind(AB1 = 3, AB2 = 4, AB3 = 5)
)

Arguments

df_in

Input data.

risk_output

Name of output file.

output_prefix

Folder location to store table file.

event_size

Event size(s) to calculate risk for.

asc_bias_list

Ascertainment bias(es) to calculate risk for, must be named.

Value

Creates, and writes to file, a table showing estimated risk that one or more people will be infectious for the given input locations, event sizes and ascertainment biases.

Examples

if (FALSE) {
Canada <- LoadData("LoadCanada")
create_c19r_data(Canada)
}